Anubhab/resq-ai-system/backend-hub/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
python-socketio==5.10.0
httpx==0.25.2
asyncio==3.4.3
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
pymongo==4.6.0
motor==3.3.2